scraper-worker/pyproject.toml
[project]
name = "scraper-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "requests>=2.32.5",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
]
